moved database and game logic to services
