+++ backend/services/matchmaking.js
@@ -2,7 +2,7 @@
 
 const shortid = require('shortid');
 const { log, DEBUG, SUCCESS, ERROR, WARNING } = require('../logging');
-const super_ttt = require("../super_ttt");
+const super_ttt = require("./super_ttt");
 const {
     X, 
     O,
